lndmanage: interactive mode when run without arguments
This commit adds an interactive mode where the user is prompted for
entering the commands available for lndmanage. This has the advantage
that the network graph has to be read only once when entering the
interactive mode and makes subsequent execution much smoother.

If lndmanage is run with arguments the previous behavior is not changed,
such that it can be part of shell scripts.