rebalancer: Fix bug where budget is ignored, remove DEFAULT_MAX_FEE_RATE
This fixes a bug where the rebalancer takes the max of the two budget options, instead of the min.
Also removes the default fee rate, since it's cumbersome to keep aligning both when one just wants to set the budget.

Also cleans up a bit; since the budget and the fee rate are correlated,
it's unnecessary to keep state for both.